add mean

+++ src/localization/likelyhood.c
@@ -110,7 +110,7 @@ void displayArray(Tdoa *array, int size)
 	}
 }
 
-void computePosition(float *x, float *y, float *z, int32_t * tabTdoa, int32_t * tabRss, Tdoa *arrayTdoa12, Tdoa *arrayTdoa13, Tdoa *arrayTdoa14, Tdoa *arrayTdoa21, Tdoa *arrayTdoa23, Tdoa *arrayTdoa24, Tdoa *arrayTdoa31, Tdoa *arrayTdoa32, Tdoa *arrayTdoa34, Tdoa *arrayTdoa41, Tdoa *arrayTdoa42, Tdoa *arrayTdoa43, int size, int nbZ, int nbPtsPlan, float cubeSize, float realZ)
+void computePosition(float *x, float *y, float *z, float * tabTdoa, int32_t * tabRss, Tdoa *arrayTdoa12, Tdoa *arrayTdoa13, Tdoa *arrayTdoa14, Tdoa *arrayTdoa21, Tdoa *arrayTdoa23, Tdoa *arrayTdoa24, Tdoa *arrayTdoa31, Tdoa *arrayTdoa32, Tdoa *arrayTdoa34, Tdoa *arrayTdoa41, Tdoa *arrayTdoa42, Tdoa *arrayTdoa43, int size, int nbZ, int nbPtsPlan, float cubeSize, float realZ)
 {
 	//float tdoa12 = 0.0, tdoa13 = 0.0, tdoa14 = 0.0 ;
 	Likelyhood *arrayLikelyhood = NULL ;
@@ -138,7 +138,7 @@ void computePosition(float *x, float *y, float *z, int32_t * tabTdoa, int32_t * 
 	}
 	printf("indmin=%d\tindmax=%d\n",indMinRss+1,indMaxRss+1);
 	for(i=0;i<4;i++){
-		printf("tdoa[%d]=%d\n",i+1,tabTdoa[i]);
+		printf("tdoa[%d]=%f\n",i+1,tabTdoa[i]);
 	}
 	
 	/*if(tdoa2!=0)
